Revert "fix /api/proxy-box-corners 2"
This reverts commit e77505b468d816eee2435317652682968db0b3d4.

+++ api/express/api/detect.js
@@ -71,7 +71,9 @@ module.exports.setApp = function (app, client) {
         )
         res.status(500).send("Error retrieving box corners")
       } finally {
-        fs.unlink(req.file.path, () => {})
+        if (req.file) {
+          fs.unlink(req.file.path, () => {})
+        }
       }
     }
   )
